+++ eslint.config.mjs
@@ -2,12 +2,13 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable no-underscore-dangle */
 /* eslint-disable import/no-extraneous-dependencies */
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+
 import { FlatCompat } from "@eslint/eslintrc";
 import js from "@eslint/js";
 import consistentImport from "@fnando/eslint-plugin-consistent-import";
 import globals from "globals";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -96,6 +97,7 @@ export default [
       "@typescript-eslint/no-floating-promises": "off",
       "object-curly-newline": "off",
       "no-plusplus": "off",
+      "no-restricted-exports": "off",
 
       // Add these rules to match Prettier config and make sure we use double quotes
       quotes: ["error", "double"],
