+++ src/components/screens/AddAssetScreen/types.ts
@@ -1,0 +1,7 @@
+export type FormattedSearchAssetRecord = {
+  assetCode: string;
+  domain: string;
+  hasTrustline: boolean;
+  issuer: string;
+  isNative: boolean;
+};
