+++ src/helpers/formatAmount.ts
@@ -125,3 +125,20 @@ export const formatPercentageAmount = (
   // BigNumber already includes the negative sign in formattedNumber
   return `${formattedNumber}%`;
 };
+
+export const stroopToXlm = (
+  stroops: BigNumber | string | number,
+): BigNumber => {
+  if (stroops instanceof BigNumber) {
+    return stroops.dividedBy(1e7);
+  }
+  return new BigNumber(Number(stroops) / 1e7);
+};
+
+export const xlmToStroop = (lumens: BigNumber | string): BigNumber => {
+  if (lumens instanceof BigNumber) {
+    return lumens.times(1e7);
+  }
+  // round to nearest stroop
+  return new BigNumber(Math.round(Number(lumens) * 1e7));
+};
